+++ src/lib/index.ts
@@ -28,9 +28,8 @@ type QuantifiedTuple<T, N extends number, A extends T[] = []> =
     : [];
 
 
-export type Subscriber<T, N extends number = 0> = IsPos<N> extends true
-  ? (value: T, ...prevs: QuantifiedTuple<T, N>) => void
-  : (value: T                                 ) => void;
+export type Subscriber<T, N extends number = 0> =
+  (value: T, ...prevs: QuantifiedTuple<T, N>) => void;
 
 
 export type StartStopNotifier<T> = (
@@ -40,7 +39,7 @@ export type StartStopNotifier<T> = (
 
 
 // stolen from svelte/store's private types "SubscribeInvalidateTuple<T>" :D
-type SITuple<T, N extends number> = [Subscriber<T, N>, Invalidator<T>];
+type SITuple<T> = [Subscriber<T, number>, Invalidator<T>];
 
 
 export type Serializer<T> = {
@@ -107,7 +106,7 @@ export interface BetterReadable<T, N extends number = 0> {
   serializer  : Serializer<T> | undefined;
   isPersistent: boolean;
 
-  previous: QuantifiedTuple<BetterReadable<T, N>, N>;
+  previous: QuantifiedTuple<BetterReadable<T, 0>, N>;
 }
 
 
@@ -185,7 +184,7 @@ export function writable<
   let present = initial as AT;
 
   const previous    = []        as BetterWritable<AT>[];
-  const subscribers = new Set() as Set<SITuple<AT, 0>>;
+  const subscribers = new Set() as Set<SITuple<AT>>;
 
   const key          = configs.key;
   const isEqual      = configs.isEqual      ?? EQ    as (a: AT, b: AT) => boolean;
@@ -352,7 +351,7 @@ export function writable<
     run : Subscriber <AT, N>,
     inv?: Invalidator<AT>,
   ) {
-    const sub: SITuple<AT, 0> = [run as any, inv ?? NOP];
+    const sub: SITuple<AT> = [run, inv ?? NOP];
 
     // add the subscriber
     subscribers.add(sub);