+++ ChangeLog.md
@@ -7,6 +7,10 @@
 1. `trackers` no longer exists. Use `previous` instead.
 1. `previous` now returns the tuple of readable stores instead of the tuple of previous value.
 
+### Changes
+
+1. Some internal type refactoring.
+
 
 ## 0.4.1